+++ src/main/java/ru/skypro/homework/dto/comment/CommentDTO.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class CommentDTO {
-    private int author;
+    private Long author;
     private String authorImage;
     private String authorFirstName;
     private int createdAt;
-    private int pk;
+    private Long pk;
     private String text;
 
 }
